refactor: change type of page count from int to long

+++ src/main/java/com/octiumsoft/commons/domain/model/PagedResult.java
@@ -8,12 +8,12 @@ import java.util.Collections;
 public class PagedResult<E> {
     private final int pageNumber;
     private final int pageSize;
-    private final int pageCount;
+    private final long pageCount;
     private Collection<E> items;
 
     public PagedResult(@Min(0) int pageNumber,
                        @Min(1) int pageSize,
-                       @Min(0) int pageCount,
+                       @Min(0) long pageCount,
                        @NotNull Collection<E> items) {
         this.pageNumber = pageNumber;
         this.pageSize = pageSize;
